move runtime_util tests out of server_util_test
