+++ lib/tests/streamlit/web/server/server_test.py
@@ -43,7 +43,7 @@ from streamlit.web.server.server import (
     Server,
     start_listening,
 )
-from .message_mocks import create_dataframe_msg
+from tests.streamlit.message_mocks import create_dataframe_msg
 from .server_test_case import ServerTestCase
 
 LOGGER = get_logger(__name__)
