+++ lib/tests/streamlit/web/server/server_util_test.py
@@ -22,12 +22,7 @@ from parameterized import parameterized
 
 import streamlit.web.server.server_util as server_util
 from streamlit import config
-from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
-from streamlit.runtime import runtime_util
-from streamlit.runtime.runtime_util import is_cacheable_msg, serialize_forward_msg
 from tests import testutil
-from tests.testutil import patch_config_options
-from .message_mocks import create_dataframe_msg
 
 
 class ServerUtilTest(unittest.TestCase):
@@ -51,38 +46,6 @@ class ServerUtilTest(unittest.TestCase):
         ):
             self.assertTrue(
                 server_util.is_url_from_allowed_origins("browser.server.address")
-            )
-
-    def test_should_cache_msg(self):
-        """Test server_util.should_cache_msg"""
-        with patch_config_options({"global.minCachedMessageSize": 0}):
-            self.assertTrue(is_cacheable_msg(create_dataframe_msg([1, 2, 3])))
-
-        with patch_config_options({"global.minCachedMessageSize": 1000}):
-            self.assertFalse(is_cacheable_msg(create_dataframe_msg([1, 2, 3])))
-
-    def test_should_limit_msg_size(self):
-        max_message_size_mb = 50
-
-        runtime_util._max_message_size_bytes = None  # Reset cached value
-        with patch_config_options({"server.maxMessageSize": max_message_size_mb}):
-            # Set up a larger than limit ForwardMsg string
-            large_msg = create_dataframe_msg([1, 2, 3])
-            large_msg.delta.new_element.markdown.body = (
-                "X" * (max_message_size_mb + 10) * 1000 * 1000
-            )
-            # Create a copy, since serialize_forward_msg modifies the original proto
-            large_msg_copy = ForwardMsg()
-            large_msg_copy.CopyFrom(large_msg)
-            deserialized_msg = ForwardMsg()
-            deserialized_msg.ParseFromString(serialize_forward_msg(large_msg_copy))
-
-            # The metadata should be the same, but contents should be replaced
-            self.assertEqual(deserialized_msg.metadata, large_msg.metadata)
-            self.assertNotEqual(deserialized_msg, large_msg)
-            self.assertTrue(
-                "exceeds the message size limit"
-                in deserialized_msg.delta.new_element.exception.message
             )
 
     @parameterized.expand(
